fix: cast loop counter to uint in HC-128 key expansion
Fixes compilation error CS0266 in Hc128Core.cs line 118.

The loop counter 'i' is an int, but when added to uint values in the
expression, it causes an implicit conversion to long which cannot be
implicitly converted back to uint.

Solution: Explicitly cast 'i' to uint in the key expansion formula:
w[i] = F2(w[i - 2]) + w[i - 7] + F1(w[i - 15]) + w[i - 16] + (uint)i;

This matches the HC-128 specification where the round constant is treated
as an unsigned 32-bit value.

+++ src/HeroCrypt/Cryptography/Symmetric/Hc128/Hc128Core.cs
@@ -115,7 +115,7 @@ internal static class Hc128Core
             // Generate expanded key using message expansion
             for (var i = 16; i < 1280; i++)
             {
-                w[i] = F2(w[i - 2]) + w[i - 7] + F1(w[i - 15]) + w[i - 16] + i;
+                w[i] = F2(w[i - 2]) + w[i - 7] + F1(w[i - 15]) + w[i - 16] + (uint)i;
             }
 
             // Initialize P and Q tables
